@@ -1,4 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// @dart=2.12
 
 part of 'api.graphql.dart';
 
@@ -6,9 +7,8 @@ part of 'api.graphql.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Signal$MutationRoot _$Signal$MutationRootFromJson(Map<String, dynamic> json) {
-  return Signal$MutationRoot()..signal = json['signal'] as bool;
-}
+Signal$MutationRoot _$Signal$MutationRootFromJson(Map<String, dynamic> json) =>
+    Signal$MutationRoot()..signal = json['signal'] as bool;
 
 Map<String, dynamic> _$Signal$MutationRootToJson(
         Signal$MutationRoot instance) =>
@@ -18,11 +18,10 @@ Map<String, dynamic> _$Signal$MutationRootToJson(
 
 CreateSession$MutationRoot$UserSession
     _$CreateSession$MutationRoot$UserSessionFromJson(
-        Map<String, dynamic> json) {
-  return CreateSession$MutationRoot$UserSession()
-    ..userId = json['userId'] as String
-    ..token = json['token'] as String?;
-}
+            Map<String, dynamic> json) =>
+        CreateSession$MutationRoot$UserSession()
+          ..userId = json['userId'] as String
+          ..token = json['token'] as String?;
 
 Map<String, dynamic> _$CreateSession$MutationRoot$UserSessionToJson(
         CreateSession$MutationRoot$UserSession instance) =>
@@ -32,11 +31,10 @@ Map<String, dynamic> _$CreateSession$MutationRoot$UserSessionToJson(
     };
 
 CreateSession$MutationRoot _$CreateSession$MutationRootFromJson(
-    Map<String, dynamic> json) {
-  return CreateSession$MutationRoot()
-    ..createSessionId = CreateSession$MutationRoot$UserSession.fromJson(
-        json['createSessionId'] as Map<String, dynamic>);
-}
+        Map<String, dynamic> json) =>
+    CreateSession$MutationRoot()
+      ..createSessionId = CreateSession$MutationRoot$UserSession.fromJson(
+          json['createSessionId'] as Map<String, dynamic>);
 
 Map<String, dynamic> _$CreateSession$MutationRootToJson(
         CreateSession$MutationRoot instance) =>
@@ -45,11 +43,10 @@ Map<String, dynamic> _$CreateSession$MutationRootToJson(
     };
 
 Signals$SubscriptionRoot$Signal _$Signals$SubscriptionRoot$SignalFromJson(
-    Map<String, dynamic> json) {
-  return Signals$SubscriptionRoot$Signal()
-    ..payload = json['payload'] as String
-    ..peerId = json['peerId'] as String;
-}
+        Map<String, dynamic> json) =>
+    Signals$SubscriptionRoot$Signal()
+      ..payload = json['payload'] as String
+      ..peerId = json['peerId'] as String;
 
 Map<String, dynamic> _$Signals$SubscriptionRoot$SignalToJson(
         Signals$SubscriptionRoot$Signal instance) =>
@@ -59,11 +56,10 @@ Map<String, dynamic> _$Signals$SubscriptionRoot$SignalToJson(
     };
 
 Signals$SubscriptionRoot _$Signals$SubscriptionRootFromJson(
-    Map<String, dynamic> json) {
-  return Signals$SubscriptionRoot()
-    ..signals = Signals$SubscriptionRoot$Signal.fromJson(
-        json['signals'] as Map<String, dynamic>);
-}
+        Map<String, dynamic> json) =>
+    Signals$SubscriptionRoot()
+      ..signals = Signals$SubscriptionRoot$Signal.fromJson(
+          json['signals'] as Map<String, dynamic>);
 
 Map<String, dynamic> _$Signals$SubscriptionRootToJson(
         Signals$SubscriptionRoot instance) =>
@@ -72,10 +68,10 @@ Map<String, dynamic> _$Signals$SubscriptionRootToJson(
     };
 
 Room$SubscriptionRoot$Room _$Room$SubscriptionRoot$RoomFromJson(
-    Map<String, dynamic> json) {
-  return Room$SubscriptionRoot$Room()
-    ..users = (json['users'] as List<dynamic>).map((e) => e as String).toList();
-}
+        Map<String, dynamic> json) =>
+    Room$SubscriptionRoot$Room()
+      ..users =
+          (json['users'] as List<dynamic>).map((e) => e as String).toList();
 
 Map<String, dynamic> _$Room$SubscriptionRoot$RoomToJson(
         Room$SubscriptionRoot$Room instance) =>
@@ -84,11 +80,10 @@ Map<String, dynamic> _$Room$SubscriptionRoot$RoomToJson(
     };
 
 Room$SubscriptionRoot _$Room$SubscriptionRootFromJson(
-    Map<String, dynamic> json) {
-  return Room$SubscriptionRoot()
-    ..room = Room$SubscriptionRoot$Room.fromJson(
-        json['room'] as Map<String, dynamic>);
-}
+        Map<String, dynamic> json) =>
+    Room$SubscriptionRoot()
+      ..room = Room$SubscriptionRoot$Room.fromJson(
+          json['room'] as Map<String, dynamic>);
 
 Map<String, dynamic> _$Room$SubscriptionRootToJson(
         Room$SubscriptionRoot instance) =>
@@ -96,12 +91,11 @@ Map<String, dynamic> _$Room$SubscriptionRootToJson(
       'room': instance.room.toJson(),
     };
 
-SignalArguments _$SignalArgumentsFromJson(Map<String, dynamic> json) {
-  return SignalArguments(
-    peerId: json['peerId'] as String,
-    signal: json['signal'] as String,
-  );
-}
+SignalArguments _$SignalArgumentsFromJson(Map<String, dynamic> json) =>
+    SignalArguments(
+      peerId: json['peerId'] as String,
+      signal: json['signal'] as String,
+    );
 
 Map<String, dynamic> _$SignalArgumentsToJson(SignalArguments instance) =>
     <String, dynamic>{
@@ -109,11 +103,10 @@ Map<String, dynamic> _$SignalArgumentsToJson(SignalArguments instance) =>
       'signal': instance.signal,
     };
 
-RoomArguments _$RoomArgumentsFromJson(Map<String, dynamic> json) {
-  return RoomArguments(
-    roomId: json['roomId'] as String,
-  );
-}
+RoomArguments _$RoomArgumentsFromJson(Map<String, dynamic> json) =>
+    RoomArguments(
+      roomId: json['roomId'] as String,
+    );
 
 Map<String, dynamic> _$RoomArgumentsToJson(RoomArguments instance) =>
     <String, dynamic>{
