@@ -1,5 +1,7 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'adding_node_state.dart';
 
@@ -10,25 +12,7 @@ part of 'adding_node_state.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$AddingConnectionStateTearOff {
-  const _$AddingConnectionStateTearOff();
-
-  _None<N> none<N extends NodeData>() {
-    return _None<N>();
-  }
-
-  _Added<N> addedInput<N extends NodeData>(Port<N> input) {
-    return _Added<N>(
-      input,
-    );
-  }
-}
-
-/// @nodoc
-const $AddingConnectionState = _$AddingConnectionStateTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$AddingConnectionState<N extends NodeData> {
@@ -36,6 +20,12 @@ mixin _$AddingConnectionState<N extends NodeData> {
   TResult when<TResult extends Object?>({
     required TResult Function() none,
     required TResult Function(Port<N> input) addedInput,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function(Port<N> input)? addedInput,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -49,6 +39,12 @@ mixin _$AddingConnectionState<N extends NodeData> {
   TResult map<TResult extends Object?>({
     required TResult Function(_None<N> value) none,
     required TResult Function(_Added<N> value) addedInput,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(_None<N> value)? none,
+    TResult Function(_Added<N> value)? addedInput,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -78,23 +74,24 @@ class _$AddingConnectionStateCopyWithImpl<N extends NodeData, $Res>
 }
 
 /// @nodoc
-abstract class _$NoneCopyWith<N extends NodeData, $Res> {
-  factory _$NoneCopyWith(_None<N> value, $Res Function(_None<N>) then) =
-      __$NoneCopyWithImpl<N, $Res>;
+abstract class _$$_NoneCopyWith<N extends NodeData, $Res> {
+  factory _$$_NoneCopyWith(_$_None<N> value, $Res Function(_$_None<N>) then) =
+      __$$_NoneCopyWithImpl<N, $Res>;
 }
 
 /// @nodoc
-class __$NoneCopyWithImpl<N extends NodeData, $Res>
+class __$$_NoneCopyWithImpl<N extends NodeData, $Res>
     extends _$AddingConnectionStateCopyWithImpl<N, $Res>
-    implements _$NoneCopyWith<N, $Res> {
-  __$NoneCopyWithImpl(_None<N> _value, $Res Function(_None<N>) _then)
-      : super(_value, (v) => _then(v as _None<N>));
+    implements _$$_NoneCopyWith<N, $Res> {
+  __$$_NoneCopyWithImpl(_$_None<N> _value, $Res Function(_$_None<N>) _then)
+      : super(_value, (v) => _then(v as _$_None<N>));
 
   @override
-  _None<N> get _value => super._value as _None<N>;
+  _$_None<N> get _value => super._value as _$_None<N>;
 }
 
 /// @nodoc
+
 class _$_None<N extends NodeData> extends _None<N>
     with DiagnosticableTreeMixin {
   const _$_None() : super._();
@@ -108,12 +105,13 @@ class _$_None<N extends NodeData> extends _None<N>
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('type', 'AddingConnectionState<$N>.none'));
+        .add(DiagnosticsProperty('type', 'AddingConnectionState<$N>.none'));
   }
 
   @override
   bool operator ==(dynamic other) {
-    return identical(this, other) || (other is _None<N>);
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is _$_None<N>);
   }
 
   @override
@@ -126,6 +124,15 @@ class _$_None<N extends NodeData> extends _None<N>
     required TResult Function(Port<N> input) addedInput,
   }) {
     return none();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function(Port<N> input)? addedInput,
+  }) {
+    return none?.call();
   }
 
   @override
@@ -152,6 +159,15 @@ class _$_None<N extends NodeData> extends _None<N>
 
   @override
   @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(_None<N> value)? none,
+    TResult Function(_Added<N> value)? addedInput,
+  }) {
+    return none?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(_None<N> value)? none,
     TResult Function(_Added<N> value)? addedInput,
@@ -170,27 +186,28 @@ abstract class _None<N extends NodeData> extends AddingConnectionState<N> {
 }
 
 /// @nodoc
-abstract class _$AddedCopyWith<N extends NodeData, $Res> {
-  factory _$AddedCopyWith(_Added<N> value, $Res Function(_Added<N>) then) =
-      __$AddedCopyWithImpl<N, $Res>;
+abstract class _$$_AddedCopyWith<N extends NodeData, $Res> {
+  factory _$$_AddedCopyWith(
+          _$_Added<N> value, $Res Function(_$_Added<N>) then) =
+      __$$_AddedCopyWithImpl<N, $Res>;
   $Res call({Port<N> input});
 }
 
 /// @nodoc
-class __$AddedCopyWithImpl<N extends NodeData, $Res>
+class __$$_AddedCopyWithImpl<N extends NodeData, $Res>
     extends _$AddingConnectionStateCopyWithImpl<N, $Res>
-    implements _$AddedCopyWith<N, $Res> {
-  __$AddedCopyWithImpl(_Added<N> _value, $Res Function(_Added<N>) _then)
-      : super(_value, (v) => _then(v as _Added<N>));
+    implements _$$_AddedCopyWith<N, $Res> {
+  __$$_AddedCopyWithImpl(_$_Added<N> _value, $Res Function(_$_Added<N>) _then)
+      : super(_value, (v) => _then(v as _$_Added<N>));
 
   @override
-  _Added<N> get _value => super._value as _Added<N>;
+  _$_Added<N> get _value => super._value as _$_Added<N>;
 
   @override
   $Res call({
     Object? input = freezed,
   }) {
-    return _then(_Added<N>(
+    return _then(_$_Added<N>(
       input == freezed
           ? _value.input
           : input // ignore: cast_nullable_to_non_nullable
@@ -200,6 +217,7 @@ class __$AddedCopyWithImpl<N extends NodeData, $Res>
 }
 
 /// @nodoc
+
 class _$_Added<N extends NodeData> extends _Added<N>
     with DiagnosticableTreeMixin {
   const _$_Added(this.input) : super._();
@@ -223,19 +241,19 @@ class _$_Added<N extends NodeData> extends _Added<N>
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other is _Added<N> &&
-            (identical(other.input, input) ||
-                const DeepCollectionEquality().equals(other.input, input)));
+        (other.runtimeType == runtimeType &&
+            other is _$_Added<N> &&
+            const DeepCollectionEquality().equals(other.input, input));
   }
 
   @override
   int get hashCode =>
-      runtimeType.hashCode ^ const DeepCollectionEquality().hash(input);
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(input));
 
   @JsonKey(ignore: true)
   @override
-  _$AddedCopyWith<N, _Added<N>> get copyWith =>
-      __$AddedCopyWithImpl<N, _Added<N>>(this, _$identity);
+  _$$_AddedCopyWith<N, _$_Added<N>> get copyWith =>
+      __$$_AddedCopyWithImpl<N, _$_Added<N>>(this, _$identity);
 
   @override
   @optionalTypeArgs
@@ -244,6 +262,15 @@ class _$_Added<N extends NodeData> extends _Added<N>
     required TResult Function(Port<N> input) addedInput,
   }) {
     return addedInput(input);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? none,
+    TResult Function(Port<N> input)? addedInput,
+  }) {
+    return addedInput?.call(input);
   }
 
   @override
@@ -270,6 +297,15 @@ class _$_Added<N extends NodeData> extends _Added<N>
 
   @override
   @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(_None<N> value)? none,
+    TResult Function(_Added<N> value)? addedInput,
+  }) {
+    return addedInput?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(_None<N> value)? none,
     TResult Function(_Added<N> value)? addedInput,
@@ -283,11 +319,11 @@ class _$_Added<N extends NodeData> extends _Added<N>
 }
 
 abstract class _Added<N extends NodeData> extends AddingConnectionState<N> {
-  const factory _Added(Port<N> input) = _$_Added<N>;
+  const factory _Added(final Port<N> input) = _$_Added<N>;
   const _Added._() : super._();
 
   Port<N> get input => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  _$AddedCopyWith<N, _Added<N>> get copyWith =>
+  _$$_AddedCopyWith<N, _$_Added<N>> get copyWith =>
       throw _privateConstructorUsedError;
 }
