@@ -6,7 +6,7 @@ part of 'input_layer.dart';
 // StoreGenerator
 // **************************************************************************
 
-// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
+// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
 
 mixin _$Input on _Input, Store {
   Computed<Tensor>? _$tensorComputed;
@@ -16,7 +16,7 @@ mixin _$Input on _Input, Store {
           Computed<Tensor>(() => super.tensor, name: '_Input.tensor'))
       .value;
 
-  final _$dtypeAtom = Atom(name: '_Input.dtype');
+  late final _$dtypeAtom = Atom(name: '_Input.dtype', context: context);
 
   @override
   DType get dtype {
@@ -31,7 +31,7 @@ mixin _$Input on _Input, Store {
     });
   }
 
-  final _$shapeAtom = Atom(name: '_Input.shape');
+  late final _$shapeAtom = Atom(name: '_Input.shape', context: context);
 
   @override
   List<int> get shape {
