@@ -6,10 +6,11 @@ part of 'tasks_store.dart';
 // StoreGenerator
 // **************************************************************************
 
-// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
+// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
 
 mixin _$TasksStore on _TasksStore, Store {
-  final _$_TasksStoreActionController = ActionController(name: '_TasksStore');
+  late final _$_TasksStoreActionController =
+      ActionController(name: '_TasksStore', context: context);
 
   @override
   void addTask() {
@@ -23,11 +24,22 @@ mixin _$TasksStore on _TasksStore, Store {
   }
 
   @override
-  bool removeTask(Task task) {
+  bool removeTask(Task task, {bool removeChildren = true}) {
     final _$actionInfo = _$_TasksStoreActionController.startAction(
         name: '_TasksStore.removeTask');
     try {
-      return super.removeTask(task);
+      return super.removeTask(task, removeChildren: removeChildren);
+    } finally {
+      _$_TasksStoreActionController.endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  Task addChildTask(Task parentTask) {
+    final _$actionInfo = _$_TasksStoreActionController.startAction(
+        name: '_TasksStore.addChildTask');
+    try {
+      return super.addChildTask(parentTask);
     } finally {
       _$_TasksStoreActionController.endAction(_$actionInfo);
     }
